Move StreetPreview code out into its own class
Also includes the correction of an assert which was bogus.
