+++ app/src/main/java/org/scottishtecharmy/soundscape/geoengine/GeoEngine.kt
@@ -37,10 +37,8 @@ import org.scottishtecharmy.soundscape.geoengine.mvttranslation.InterpolatedPoin
 import org.scottishtecharmy.soundscape.geoengine.mvttranslation.vectorTileToGeoJson
 import org.scottishtecharmy.soundscape.geoengine.utils.FeatureTree
 import org.scottishtecharmy.soundscape.geoengine.utils.RelativeDirections
-import org.scottishtecharmy.soundscape.geoengine.utils.RoadDirectionAtIntersection
 import org.scottishtecharmy.soundscape.geoengine.utils.TileGrid
 import org.scottishtecharmy.soundscape.geoengine.utils.TileGrid.Companion.getTileGrid
-import org.scottishtecharmy.soundscape.geoengine.utils.bearingFromTwoPoints
 import org.scottishtecharmy.soundscape.geoengine.utils.checkIntersection
 import org.scottishtecharmy.soundscape.geoengine.utils.cleanTileGeoJSON
 import org.scottishtecharmy.soundscape.geoengine.utils.deduplicateFeatureCollection
@@ -48,7 +46,6 @@ import org.scottishtecharmy.soundscape.geoengine.utils.distance
 import org.scottishtecharmy.soundscape.geoengine.utils.distanceToPolygon
 import org.scottishtecharmy.soundscape.geoengine.utils.getCompassLabelFacingDirection
 import org.scottishtecharmy.soundscape.geoengine.utils.getCompassLabelFacingDirectionAlong
-import org.scottishtecharmy.soundscape.geoengine.utils.getDirectionAtIntersection
 import org.scottishtecharmy.soundscape.geoengine.utils.getFeatureNearestPoint
 import org.scottishtecharmy.soundscape.geoengine.utils.getFovIntersectionFeatureCollection
 import org.scottishtecharmy.soundscape.geoengine.utils.getFovRoadsFeatureCollection
@@ -66,12 +63,8 @@ import org.scottishtecharmy.soundscape.geoengine.utils.processTileFeatureCollect
 import org.scottishtecharmy.soundscape.geoengine.utils.processTileString
 import org.scottishtecharmy.soundscape.geoengine.utils.removeDuplicateOsmIds
 import org.scottishtecharmy.soundscape.geoengine.utils.sortedByDistanceTo
-import org.scottishtecharmy.soundscape.geoengine.utils.distanceToLineString
-import org.scottishtecharmy.soundscape.geoengine.utils.splitRoadAtNode
-import org.scottishtecharmy.soundscape.geoengine.utils.splitRoadByIntersection
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.Feature
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.FeatureCollection
-import org.scottishtecharmy.soundscape.geojsonparser.geojson.LineString
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.LngLatAlt
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.Point
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.Polygon
@@ -88,7 +81,6 @@ import org.scottishtecharmy.soundscape.utils.getCurrentLocale
 import retrofit2.awaitResponse
 import java.util.Locale
 import kotlin.coroutines.cancellation.CancellationException
-import kotlin.math.abs
 import kotlin.time.TimeSource
 
 data class PositionedString(
@@ -110,8 +102,8 @@ class GeoEngine {
     // HTTP connection to soundscape-backend or protomaps tile server
     private lateinit var tileClient: TileClient
 
-    private lateinit var locationProvider: LocationProvider
-    private lateinit var directionProvider: DirectionProvider
+    internal lateinit var locationProvider : LocationProvider
+    private lateinit var directionProvider : DirectionProvider
 
     // Resource string locale configuration
     private lateinit var configLocale: Locale
@@ -125,6 +117,8 @@ class GeoEngine {
     private var featureTrees = Array(Fc.MAX_COLLECTION_ID.id) { FeatureTree(null) }
     @OptIn(ExperimentalCoroutinesApi::class, DelicateCoroutinesApi::class)
     private val treeContext = newSingleThreadContext("TreeContext")
+
+    private val streetPreview = StreetPreview()
 
     @Subscribe
     fun onActivityTransitionEvent(event: ActivityTransitionEvent) {
@@ -164,7 +158,7 @@ class GeoEngine {
         locationProvider = newLocationProvider
         directionProvider = newDirectionProvider
 
-        previewState = PreviewState.INITIAL
+        streetPreview.start()
 
         startTileGridService(soundscapeService)
 
@@ -1437,10 +1431,10 @@ class GeoEngine {
         MAX_COLLECTION_ID(8),
     }
 
-    private fun getGridFeatureCollection(id: Int,
-                                         location: LngLatAlt = LngLatAlt(),
-                                         distance : Double = Double.POSITIVE_INFINITY,
-                                         maxCount : Int = 0): FeatureCollection {
+    internal fun getGridFeatureCollection(id: Int,
+                                          location: LngLatAlt = LngLatAlt(),
+                                          distance : Double = Double.POSITIVE_INFINITY,
+                                          maxCount : Int = 0): FeatureCollection {
         val result = if(distance == Double.POSITIVE_INFINITY) {
             featureTrees[id].generateFeatureCollection()
         } else {
@@ -1457,20 +1451,6 @@ class GeoEngine {
         return result
     }
 
-    enum class PreviewState(val id: Int) {
-        INITIAL(0),
-        AT_NODE(1)
-    }
-
-    data class StreetPreviewChoice (
-        val heading: Double,
-        val name: String,
-        val route: List<LngLatAlt>
-    )
-
-    private var previewState = PreviewState.INITIAL
-    private var previewRoad : StreetPreviewChoice? = null
-
     fun streetPreviewGo() {
 
         val start = System.currentTimeMillis()
@@ -1483,314 +1463,12 @@ class GeoEngine {
         val heading = directionProvider.getCurrentDirection()
         // Run the code within the treeContext to protect it from changes to the trees whilst it's
         // running.
+        val engine = this
         CoroutineScope(Job()).launch(treeContext) {
-            when(previewState) {
-
-                PreviewState.INITIAL -> {
-                    // Jump to a node on the nearest road or path
-                    val roads = getGridFeatureCollection(Fc.ROADS.id, location, 200.0, 1)
-                    roads.plusAssign(getGridFeatureCollection(Fc.PATHS.id, location, 200.0, 1))
-
-                    var nearestDistance = Double.POSITIVE_INFINITY
-                    var nearestPoint = LngLatAlt()
-                    for (road in roads) {
-                        val nearestPointOnRoad = LngLatAlt()
-                        val distance = distanceToLineString(location,
-                            road.geometry as LineString,
-                            nearestPointOnRoad)
-                        if(distance < nearestDistance) {
-                            nearestDistance = distance
-                            nearestPoint = nearestPointOnRoad
-                        }
-                    }
-                    if(nearestDistance != Double.POSITIVE_INFINITY) {
-                        // We've got a location, so jump to it
-                        locationProvider.updateLocation(nearestPoint, 0.0F)
-                        previewState = PreviewState.AT_NODE
-                    }
-                }
-
-                PreviewState.AT_NODE -> {
-                    // Find which road that we're choosing based on our current heading
-                    val choices = getDirectionChoices(location)
-                    var bestIndex = -1
-                    var bestHeadingDiff = Double.POSITIVE_INFINITY
-
-                    // Find the choice with the closest heading to our own
-                    var diff: Double
-                    for((index, choice) in choices.withIndex()) {
-                        diff = abs(choice.heading - heading)
-                        if(diff < bestHeadingDiff) {
-                            bestHeadingDiff = diff
-                            bestIndex = index
-                        }
-                        Log.d(TAG, "Choice: ${choice.name} heading: ${choice.heading}")
-                    }
-                    // Check that the closest heading is close enough
-                    if(bestHeadingDiff < 30.0) {
-
-                        // We've got a road - let's head down it
-                        previewRoad = extendChoice(location, choices[bestIndex])
-                        previewRoad?.let { road ->
-                            locationProvider.updateLocation(road.route.last(), 1.0F)
-                        }
-                        previewState = PreviewState.AT_NODE
-                    }
-                }
-            }
+            streetPreview.go(location, heading, engine)
         }
         val end = System.currentTimeMillis()
         Log.d(TAG, "streetPreviewGo: ${end-start}ms")
-    }
-
-    private fun bearingOfLineFromEnd(location: LngLatAlt, line: List<LngLatAlt>) : Double {
-        var heading = Double.NaN
-        var nextPoint : LngLatAlt? = null
-        if(location == line.first()) {
-            if(line.size <= 1) {
-                Log.e(TAG, "bearingOfLineFromEnd: line too short")
-            }
-            else {
-                nextPoint = line[1]
-            }
-        } else if(location == line.last()) {
-            nextPoint = line.dropLast(1).last()
-        }
-        if(nextPoint != null) {
-            heading = bearingFromTwoPoints(
-                location.latitude,
-                location.longitude,
-                nextPoint.latitude,
-                nextPoint.longitude
-            )
-        }
-        return heading
-    }
-
-    /**
-     * extendChoice extends the road line in the choice in the direction of travel. It  also orders
-     * the points in the line so that first() is the current location and last() is one of either:
-     *  1. The next intersection (most common)
-     *  2. The end of the road
-     *  3. The edge of the tile grid
-     */
-    private fun extendUntilIntersection(newLine : List<LngLatAlt>, extendedLine : MutableList<LngLatAlt>) : Boolean {
-        var foundIntersection = false
-        for((index, point) in newLine.withIndex()) {
-            extendedLine.add(point)
-
-            if(index != 0) {
-                // Check for an intersection at this point by searching in the intersection tree
-                val intersection = getGridFeatureCollection(Fc.INTERSECTIONS.id, point, 0.5, 1)
-                if (intersection.features.isNotEmpty()) {
-                    Log.e(
-                        TAG,
-                        "Stop at intersection ${intersection.features[0].properties?.get("name")}"
-                    )
-                    foundIntersection = true
-                    break
-                }
-            }
-        }
-        return foundIntersection
-    }
-
-    private fun extendChoice(location: LngLatAlt, choice : StreetPreviewChoice) : StreetPreviewChoice {
-        var line = choice.route
-        if(location == line.last()) {
-            line = line.reversed()
-        }
-
-        val extendedLine = mutableListOf<LngLatAlt>()
-
-        while(true) {
-            // Find any intersections in the current line segment
-            val foundIntersection = extendUntilIntersection(line, extendedLine)
-
-            if (foundIntersection) {
-                return StreetPreviewChoice(choice.heading, choice.name, extendedLine)
-            }
-
-            // We've not found an intersection yet, so we've reached the end of the line
-            var foundNewRoad = false
-            if(line.size < 2) {
-                return StreetPreviewChoice(choice.heading, choice.name, extendedLine)
-            }
-            val currentPoint = line.last()
-            val previousPoint = line.dropLast(1).last()
-            val roads = getGridFeatureCollection(Fc.ROADS.id, currentPoint, 0.5, 3)
-            roads.plusAssign(getGridFeatureCollection(Fc.PATHS.id, currentPoint, 0.5, 3))
-            for (road in roads) {
-                // Find which roads the currentPoint is in
-                val roadPoints = road.geometry as LineString
-                val containsCurrentPoint = roadPoints.coordinates.any { it == currentPoint }
-                val containsPreviousPoint = roadPoints.coordinates.any { it == previousPoint }
-                if (containsCurrentPoint && containsPreviousPoint)
-                {
-                    // This is our line, so ignore it
-                } else if (containsCurrentPoint) {
-                    // This is the next line
-                    line = roadPoints.coordinates
-                    if(currentPoint == line.last()) {
-                        line = line.reversed()
-                    }
-                    foundNewRoad = true
-                    break
-                }
-            }
-            if(!foundNewRoad) {
-                // Return what we have
-                return StreetPreviewChoice(choice.heading, choice.name, extendedLine)
-            }
-        }
-    }
-
-    /**
-     * getDirectionChoices returns a List of possible choices at an intersection
-     * Each entry contains a heading, the street name and a list of points that make up the road.
-     * The app can choose the road based on the heading and then move the user along it.
-     */
-    private fun getDirectionChoices(location: LngLatAlt) : List<StreetPreviewChoice> {
-        val choices = mutableListOf<StreetPreviewChoice>()
-
-        val start = System.currentTimeMillis()
-
-        val nearestIntersection = getGridFeatureCollection(Fc.INTERSECTIONS.id, location, 1.0, 1)
-        val nearestRoads = getGridFeatureCollection(Fc.ROADS.id, location, 1.0)
-        nearestRoads.plusAssign(getGridFeatureCollection(Fc.PATHS.id, location, 1.0))
-
-        if(nearestIntersection.features.isNotEmpty()) {
-            // We're at an intersection
-            val intersectionRoadNames = getIntersectionRoadNames(nearestIntersection, nearestRoads)
-            for (road in intersectionRoadNames) {
-                val testRoadDirectionAtIntersection =
-                    getDirectionAtIntersection(nearestIntersection.features[0], road)
-
-                if (testRoadDirectionAtIntersection == RoadDirectionAtIntersection.LEADING_AND_TRAILING) {
-                    // split the road into two
-                    val roadCoordinatesSplitIntoTwo = splitRoadByIntersection(
-                        nearestIntersection.features[0],
-                        road
-                    )
-
-                    for (splitRoad in roadCoordinatesSplitIntoTwo) {
-                        road.properties?.get("name")?.let { choice ->
-                            val line = splitRoad.geometry as LineString
-                            val heading = bearingOfLineFromEnd(location, line.coordinates)
-                            choices.add(
-                                StreetPreviewChoice(
-                                    heading,
-                                    choice.toString(),
-                                    line.coordinates
-                                )
-                            )
-                        }
-                    }
-                } else if (testRoadDirectionAtIntersection != RoadDirectionAtIntersection.NONE) {
-                    road.properties?.get("name")?.let { choice ->
-                        val line = road.geometry as LineString
-                        val heading = bearingOfLineFromEnd(location, line.coordinates)
-                        choices.add(
-                            StreetPreviewChoice(
-                                heading,
-                                choice.toString(),
-                                line.coordinates
-                            )
-                        )
-                    }
-                }
-            }
-        }
-        else {
-            // There are a couple of reasons that we could be here:
-            //
-            //  1. We're in the middle of a road straight after starting StreetPreview
-            //  2. We reached the edge of the tile grid on the last jump - the road had no
-            //     intersections between the central bounding box and the grid edge, or it was slow
-            //     in loading up the tile grid and we jumped along the road very fast.
-            //
-            // In the first case we likely have a single road within 1m that we need to split into
-            // two, but in the second case we likely have 3 nearby roads, one of which is a short
-            // joining road between two tiles and it's also pointing in the wrong direction! We want
-            // to follow along the joining road and then use the road that it joins on to for the
-            // heading.
-            var currentRoad : Feature? = null
-            var joiningRoadOtherEnd : LngLatAlt? = null
-            var adjoiningRoads = FeatureCollection()
-
-            // Add the current road and find out if we have a short joining road
-            for(road in nearestRoads.features) {
-                val lineString = road.geometry as LineString
-                val containsCurrentPoint = lineString.coordinates.any { it == location }
-                if(containsCurrentPoint) {
-                    if(road.foreign?.containsKey("tileJoiner") != true) {
-                        adjoiningRoads.addFeature(road)
-                        currentRoad = road
-                    }
-                    else {
-                        var index = 0
-                        if(lineString.coordinates[0] == location) {
-                            index = 1
-                        }
-                        assert(lineString.coordinates[1] == location)
-                        joiningRoadOtherEnd = lineString.coordinates[index]
-                    }
-                }
-            }
-            if(joiningRoadOtherEnd != null) {
-                // Add the road at the other end of the joining road
-                for(road in nearestRoads.features) {
-                    val lineString = road.geometry as LineString
-                    val containsOtherEnd = lineString.coordinates.any { it == joiningRoadOtherEnd }
-                    if(containsOtherEnd) {
-                        if(road.foreign?.containsKey("tileJoiner") != true) {
-                            // This is the road at the other end of the joining road
-                            adjoiningRoads.addFeature(road)
-                        }
-                    }
-                }
-            }
-
-            if((adjoiningRoads.features.size == 1) && (currentRoad != null)) {
-                // We've only got one road, so split it
-                adjoiningRoads = splitRoadAtNode(location, currentRoad)
-            }
-
-            // Create the choices from the roads that we've selected
-            for (road in adjoiningRoads) {
-                val line = road.geometry as LineString
-                var lineList : List<LngLatAlt> = line.coordinates
-                if(joiningRoadOtherEnd != null) {
-                    val containsOtherEnd = line.coordinates.any { it == joiningRoadOtherEnd }
-                    if(containsOtherEnd) {
-                        // We've got the line after the joining road. Copy it and replace the end of
-                        // it with our current location
-                        val newLine: MutableList<LngLatAlt> = mutableListOf()
-                        for (point in lineList) {
-                            if (point == joiningRoadOtherEnd) {
-                                // This moves the end of the line to our current location, thus
-                                // skipping the joining line.
-                                newLine.add(location)
-                            } else
-                                newLine.add(point)
-                        }
-                        lineList = newLine
-                    }
-                }
-                val heading = bearingOfLineFromEnd(location, lineList)
-                choices.add(
-                    StreetPreviewChoice(
-                        heading,
-                        name = road.properties?.get("name").toString(),
-                        lineList
-                    )
-                )
-            }
-        }
-        val end = System.currentTimeMillis()
-        Log.d(TAG, "getDirectionChoices: ${end-start}ms")
-
-        return choices
     }
 
     companion object {
